Moving resources object to resources.js
